Render NetworkAddress indices (#87)
* Add OID-index roundtrip methods to NetworkAddress

This enables use of NetworkAddress as a table index.

* Use NetworkAddress for RFC1213-MIB::atNetAddress

Previously atNetAddress defined to be an IpAddress, whose table index
mapping was different from that of NetworkAddress.  This prevented
proper use of RFC1213-MIB::atTable instances, because:

- OID-to-symbol resolution was failing;
- Symbol-to-OID mapping result was invalid.

* Move clone() from Choice to NetworkAddress

Previously it was implemented in pyasn1.type.univ.Choice in case there
may be more Choice-based types—such as NetworkAddress—used as a table
index.  However, SMIv2 (RFC 2578) limits the SYNTAX of an OBJECT-TYPE to
be only PyASN1 “simple” types, and NetworkAddress is the only known
Choice-based type used as a table index in MIB-I, so there is little
reason to clutter PyASN1 with the one-off logic in anticipation of
something will probably never happen.

Having NetworkAddress's own clone() method also allows use of string
literals as the value, so the following invocations are all valid:

    na = NetworkAddress()
    na1234 = na.clone('1.2.3.4')
    na1234_2 = na1234.clone()
    na1234_3 = na.clone(na1234)
    na4321 = na.clone(IpAddress('4.3.2.1'))

To elaborate on simple types, SMIv2 limits the object syntax to be:

- a base type (or its refinement)
- a textual convention (or its refinement); or
- a BITS pseudo-type.

All base types descend from ASN.1 integer, octet string, or OID, all of
which are simple types.  PySNMP defines SMIv2 BITS as a subclass of
OctetString, which is again a simple type.  Finally, a SMIv2 textual
convention (RFC 2579) is simply a syntactic sugar applied on top of
either a base type a BITS type, so it is a simple type.