Refactor MIB state machine into asynchronous operations
When MIB instrumentation controller reads or writes managed objects,
it does so in transactional manner e.g. going through a sequence of
method calls for each managed object.

This commit refactors the MIB instrumentation methods being called
by the state machine to return immediately and resume once the callback
is called.

This change is a prerequisite for fully asynchronous managed objects
implementation.