Do not store incomplete USM keys and improve debug
This adds details debugging on USM initial configuration process
and runtime USM user cloning.

Besides that, this patch eliminates storing of incomplete
USM keys (in case when master/localized keys are configured
directly).

On top of that, this commit fixes a bug in USM configuration
which did not allow the same user names to be added under
different security names.